+++ test/3250-ui_option_window_open_and_close_test.dart
@@ -186,6 +186,59 @@ void main() {
 
         // Verify the MPIDOptionWidget is gone after pressing Ok
         expect(find.byType(MPIDOptionWidget), findsNothing);
+
+        final Finder mpTileWidgetWithVisibilityFinder = find.descendant(
+          of: editWidgetFinder,
+          matching: find.ancestor(
+            of: find.text('Visibility'),
+            matching: find.byType(MPTileWidget),
+          ),
+        );
+
+        expect(mpTileWidgetWithVisibilityFinder, findsOneWidget);
+
+        // check MPTileWidget background color equals
+        // theme.colorScheme.surfaceContainer, i.e., is unset.
+        final MPTileWidget mpTileWidgetVisibility = tester.widget<MPTileWidget>(
+          mpTileWidgetWithVisibilityFinder,
+        );
+
+        expect(mpTileWidgetVisibility.backgroundColor, unsetExpectedColor);
+
+        // tap the MPTileWidget and verify MPMultipleOptionWidget opens
+        await tester.tap(mpTileWidgetWithVisibilityFinder);
+        await tester.pumpAndSettle();
+
+        final Finder mpVisibilityOptionFinder = find.byType(
+          MPMultipleChoicesWidget,
+        );
+
+        expect(mpVisibilityOptionFinder, findsOneWidget);
+
+        // Click the RadioListTile to select the SET option
+        final Finder offRadioFinder = find.byKey(
+          const ValueKey(
+            'MPMultipleChoicesWidget|visibility|RadioListTile|off',
+          ),
+        );
+
+        expect(offRadioFinder, findsOneWidget);
+        await tester.ensureVisible(offRadioFinder);
+        await tester.pumpAndSettle();
+        await tester.tap(offRadioFinder);
+        await tester.pumpAndSettle();
+
+        // /// TODO
+        // // check MPTileWidget background color equals
+        // // theme.colorScheme.tertiaryFixed, i.e., is set.
+        // // final Color setExpectedColor = Theme.of(
+        // //   tester.element(editWidgetFinder),
+        // // ).colorScheme.tertiaryFixed;
+
+        // // expect(mpTileWidgetID.backgroundColor, setExpectedColor);
+
+        // Verify the MPMultipleChoicesWidget is gone after pressing Ok
+        expect(find.byType(MPMultipleChoicesWidget), findsNothing);
       },
     );
   });